APIS-4714 - Add testing pages controller spec
